2009-05-25  Ramon Novoa  <rnovoa@artica.es>
	* windows/pandora_wmi.cc, windows/pandora_wmi.h: Added functions for the
          new modules. Removed old logevent functions.

        * pandora_windows_service.h, pandora_windows_service.cc: Added support
          for data file buffering and startup delay.

        * modules/pandora_module_logevent.cc, modules/pandora_module_logevent.h:
          Rewritten to retrieve event log data using the PDH interface instead
          of WMI (was too slow).

        * modules/pandora_module_tcpcheck.cc, modules/pandora_module_tcpcheck.h,
          modules/pandora_module_regexp.cc, modules/pandora_module_regexp.h,
          modules/pandora_module_perfcounter.cc,
          modules/pandora_module_perfcounter.h,
          modules/pandora_module_freedisk_percent.cc,
          modules/pandora_module_freedisk_percent.h,
          modules/pandora_module_freememory_percent.cc,
          modules/pandora_module_freememory_percent.h: Added to repository.
          New modules, see http://openideas.info/wiki/.

        * modules/pandora_module.h, modules/pandora_module.cc,
          modules/pandora_module_factory.cc, modules/pandora_module_list.cc:
          Added support for the new modules.

        * bin/pandora_agent.conf: Included configuration examples for the new
          modules.

        * PandoraAgent.dev: Updated to compile the new modules.




git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@1707 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_agents/win32/pandora_windows_service.cc
@@ -683,6 +683,12 @@ Pandora_Windows_Service::sendXml (Pandora_Module_List *modules) {
 	string            encoding;
 	bool              saved;
 	static HANDLE     mutex = 0; 
+    ULARGE_INTEGER    free_bytes;
+    double            min_free_bytes = 0;
+	Pandora_Agent_Conf *conf = NULL;
+
+	conf = this->getConf ();
+	min_free_bytes = 1024 * atoi (conf->getValue ("temporal_min_size").c_str ());
 	
 	if (mutex == 0) {
 		mutex = CreateMutex (NULL, FALSE, NULL);
@@ -749,24 +755,76 @@ Pandora_Windows_Service::sendXml (Pandora_Module_List *modules) {
 
 	/* Only send if debug is not activated */
 	if (getPandoraDebug () == false) {
-		this->copyDataFile (tmp_filename);
-        Pandora_File::removeFile (tmp_filepath);
+		rc = this->copyDataFile (tmp_filename);
+        
+        /* Delete the file if successfully copied or not enough space available */
+        if (rc == 0 || (GetDiskFreeSpaceEx (tmp_filepath.c_str (), &free_bytes, NULL, NULL) != 0
+            && free_bytes.QuadPart < min_free_bytes)) {
+            Pandora_File::removeFile (tmp_filepath);
+	    }
+
+        /* Send any buffered data files */
+        this->sendBufferedXml (conf->getValue ("temporal"));
 	}
-	
+
 	ReleaseMutex (mutex);
+}
+
+void
+Pandora_Windows_Service::sendBufferedXml (string path) {
+    string base_path = path, file_path;
+    WIN32_FIND_DATA file_data;
+    HANDLE find;
+
+	if (base_path[base_path.length () - 1] != '\\') {
+		base_path += "\\";
+	}
+    file_path = base_path + "*.data";
+    
+    /* Search for buffered data files */
+    find = FindFirstFile(file_path.c_str (), &file_data);
+    if (find == INVALID_HANDLE_VALUE) {
+        return;
+    }
+
+    /* Send data files as long as there are no errors */
+    if (this->copyDataFile (file_data.cFileName) != 0) {
+        FindClose(find);
+        return;
+    }
+    Pandora_File::removeFile (base_path + file_data.cFileName);
+
+    while (FindNextFile(find, &file_data) != 0) {
+        if (this->copyDataFile (file_data.cFileName) != 0) {
+            FindClose(find);
+            return;
+        }
+        Pandora_File::removeFile (base_path + file_data.cFileName);
+    }
+
+    FindClose(find);
 }
 
 void
 Pandora_Windows_Service::pandora_run () {
 	Pandora_Agent_Conf  *conf = NULL;
 	string server_addr;
+    int startup_delay = 0;
 
 	pandoraDebug ("Run begin");
 	
+	conf = this->getConf ();
+ 	
+ 	/* Sleep if a startup delay was specified */
+ 	startup_delay = atoi (conf->getValue ("startup_delay").c_str ());
+ 	if (startup_delay > 0) {
+        pandoraLog ("Delaying startup %d seconds", startup_delay);
+        Sleep (startup_delay);
+    }
+
 	/* Check for configuration changes */
 	this->checkConfig ();
 
-	conf = this->getConf ();
 	server_addr = conf->getValue ("server_ip");
 
 	execution_number++;
